feat: Initial commit, working script & simple doc
Repo features the script to walk through folders and remove duplicated
files. I needed it since I had a case where I had many files with
different name, but the same contents.

If I am to continue work on this I will most certainly make it object
oriented and pythonic, as the procedural nature of the script is a bit
tough on the eyes.